Move QueryOptions to ApolloClient namespace
