+++ src/__tests__/client.ts
@@ -15,7 +15,6 @@ import type {
   FetchPolicy,
   ObservableQuery,
   Operation,
-  QueryOptions,
   TypedDocumentNode,
   WatchQueryFetchPolicy,
 } from "@apollo/client";
@@ -1757,7 +1756,10 @@ describe("client", () => {
         cache: new InMemoryCache(),
       });
       expect(() => {
-        void client.query({ query, returnPartialData: true } as QueryOptions);
+        void client.query({
+          query,
+          returnPartialData: true,
+        } as ApolloClient.QueryOptions);
       }).toThrowError(/returnPartialData/);
     });
 
@@ -1767,7 +1769,10 @@ describe("client", () => {
         cache: new InMemoryCache(),
       });
       expect(() => {
-        void client.query({ query, returnPartialData: true } as QueryOptions);
+        void client.query({
+          query,
+          returnPartialData: true,
+        } as ApolloClient.QueryOptions);
       }).toThrowError(/returnPartialData/);
     });
   });
@@ -2199,7 +2204,10 @@ describe("client", () => {
         cache: new InMemoryCache(),
       });
 
-      const options: QueryOptions = { query, fetchPolicy: "network-only" };
+      const options: ApolloClient.QueryOptions = {
+        query,
+        fetchPolicy: "network-only",
+      };
 
       // Run a query first to initialize the store
       await client.query({ query });
