+++ src/core/deprecated.ts
@@ -7,6 +7,12 @@ export type ApolloClientOptions = ApolloClient.Options;
 /** @deprecated Use `ApolloClient.DevtoolsOptions` instead */
 export type DevtoolsOptions = ApolloClient.DevtoolsOptions;
 
+/** @deprecated Use `ApolloClient.QueryOptions` instead */
+export type QueryOptions<
+  TVariables extends OperationVariables = OperationVariables,
+  TData = unknown,
+> = ApolloClient.QueryOptions<TData, TVariables>;
+
 /** @deprecated Use `ApolloClient.WatchQueryOptions` instead */
 export type WatchQueryOptions<
   TVariables extends OperationVariables = OperationVariables,
